requirements.txt
accelerate==0.20.3
appnope==0.1.4
asttokens==3.0.0
blosc==1.11.3
Caffe==0.1.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.0.4
comm==0.2.3
debugpy==1.8.15
decorator==5.2.1
diffusers==0.34.0
executing==2.2.0
filelock==3.18.0
fsspec==2025.7.0
hf-xet==1.1.7
huggingface-hub==0.34.3
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
loguru==0.7.3
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.2
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
pyzmq==27.0.1
regex==2025.7.34
requests==2.32.4
safetensors==0.6.1
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tokenizers==0.21.4
torch==2.8.0
torchvision==0.23.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.0
typer==0.9.0
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
zipp==3.23.0
